[#DCM-146]: Fix Eclipse warnings - Unnecessary cast from x to x

+++ dcm4che2/trunk/dcm4che-core/src/main/java/org/dcm4che2/io/ContentHandlerAdapter.java
@@ -219,7 +219,7 @@ public class ContentHandlerAdapter extends DefaultHandler {
             URLConnection con = url.openConnection();
             DataInputStream in = new DataInputStream(con.getInputStream());
             try {
-                int len = (int) con.getContentLength();
+                int len = con.getContentLength();
                 byte[] data = new byte[(len + 1) & ~1];
                 in.readFully(data, 0, len);
                 return data;
